fix(rich-text): unwrap pasting of single table cells (#1094)
